fix: reorganize dockerfile analysis to own module
Removes it from the top-level lib/ module as the Dockerfile analysis logic was split across two files. This change unifies it and also exports types from types.ts.
