+++ lib/dockerfile/types.ts
@@ -1,0 +1,17 @@
+export interface DockerFileAnalysis {
+  baseImage?: string;
+  dockerfilePackages: DockerFilePackages;
+  dockerfileLayers: DockerFileLayers;
+}
+
+export interface DockerFilePackages {
+  [packageName: string]: {
+    instruction: string;
+  };
+}
+
+export interface DockerFileLayers {
+  [id: string]: {
+    instruction: string;
+  };
+}
