+++ test/system/bugs/rpm-transitive-dependencies.spec.ts
@@ -1,6 +1,6 @@
 import { join as pathJoin } from "path";
 import { scan } from "../../../lib";
-import { DockerFileAnalysis } from "../../../lib/docker-file";
+import { DockerFileAnalysis } from "../../../lib/dockerfile";
 
 function getFixture(fixturePath) {
   return pathJoin(__dirname, "../../fixtures", fixturePath);
